Added product-listing.js in js folder
